Add selection of multiple Pi's
Add selection of different Pi's (4 missing BT-addresses).
Add functionality to dim and function buttons and remove unnecessary lines of code in switch-case statement for light buttons.
Fix bug in which the dim buttons used the wrong alpha value if the slider for alpha had been adjusted manually.
Clean up toasts and logging for status messages when connecting/disconnecting from a controller; these should now show in correct order and the correct amount.
Fix fatal bug which caused an unrecoverable crash if the app was exited with "connect automatically" selected and a Pi with no address was selected.

Signed-off-by: Kodemikkel <martin.krisknut@hotmail.com>